+++ tests/unit_tests/Commands/AliasesCommandTest.php
@@ -8,6 +8,8 @@ use Pantheon\Terminus\Helpers\LocalMachineHelper;
 use Robo\Config;
 use Pantheon\Terminus\Models\User;
 use Pantheon\Terminus\Session\Session;
+use Pantheon\Terminus\UnitTests\Helpers\AliasFixtures;
+use Symfony\Component\Console\Output\BufferedOutput;
 
 /**
  * Class AliasesCommandTest
@@ -17,10 +19,6 @@ use Pantheon\Terminus\Session\Session;
 class AliasesCommandTest extends CommandTestCase
 {
     /**
-     * @var string
-     */
-    protected $aliases;
-    /**
      * @var Config
      */
     protected $config;
@@ -29,17 +27,18 @@ class AliasesCommandTest extends CommandTestCase
      */
     protected $container;
     /**
-     * @var LocalMachineHelper
-     */
-    protected $local_machine_helper;
-    /**
      * @var Session
      */
     protected $session;
     /**
-     * @var User
+     * @var string
      */
-    protected $user;
+    protected $home_dir;
+    /**
+     * @var BufferedOutput
+     */
+    protected $output;
+
 
     /**
      * @inheritdoc
@@ -48,33 +47,44 @@ class AliasesCommandTest extends CommandTestCase
     {
         parent::setUp();
 
-        $this->aliases = '//Aliases';
+        $this->fixtures = new AliasFixtures();
+        $this->output = new BufferedOutput();
+
+        $this->home_dir = realpath($this->fixtures->mktmpdir());
+
         $this->session = $this->getMockBuilder(Session::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $this->user = $this->getMockBuilder(User::class)
             ->disableOriginalConstructor()
             ->getMock();
         $this->container = $this->getMockBuilder(Container::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->local_machine_helper = $this->getMockBuilder(LocalMachineHelper::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->session->expects($this->once())
-            ->method('getUser')
-            ->with()
-            ->willReturn($this->user);
-        $this->user->expects($this->once())
-            ->method('getAliases')
-            ->with()
-            ->willReturn($this->aliases);
+        $this->getConfig()->method('get')
+            ->with($this->equalTo('user_home'))
+            ->willReturn($this->home_dir);
+        $this->sites->method('fetch')
+            ->willReturn(null);
+        $this->sites->method('ids')
+            ->willReturn([$this->site->id]);
+        $this->sites->method('serialize')
+            ->willReturn([$this->site->id => ['id' => $this->site->id, 'name' => 'site1']]);
+        $this->site->method('get')
+            ->willReturn('site1');
 
         $this->command = new AliasesCommand($this->getConfig());
         $this->command->setLogger($this->logger);
         $this->command->setSession($this->session);
         $this->command->setContainer($this->container);
+        $this->command->setConfig($this->config);
+        $this->command->setSites($this->sites);
+        $this->command->setOutput($this->output);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function tearDown()
+    {
+        $this->fixtures->cleanup();
     }
 
     /**
@@ -82,57 +92,75 @@ class AliasesCommandTest extends CommandTestCase
      */
     public function testAliases()
     {
-        $default_location = '~/.drush/pantheon.aliases.drushrc.php';
-
-        $this->container->expects($this->once())
-            ->method('get')
-            ->with($this->equalTo(LocalMachineHelper::class))
-            ->willReturn($this->local_machine_helper);
-        $this->local_machine_helper->expects($this->once())
-            ->method('writeFile')
-            ->with(
-                $this->equalTo($default_location),
-                $this->equalTo($this->aliases)
-            );
-        $this->logger->expects($this->once())
+        $this->logger->expects($this->at(0))
             ->method('log')
             ->with(
                 $this->equalTo('notice'),
-                $this->equalTo('Aliases file written to {location}.'),
-                $this->equalTo(['location' => $default_location,])
+                $this->equalTo('Fetching site information to build Drush aliases...')
+            );
+        $this->logger->expects($this->at(1))
+            ->method('log')
+            ->with(
+                $this->equalTo('notice'),
+                $this->equalTo('{count} sites found.'),
+                $this->equalTo(['count' => 1])
+            );
+        $this->logger->expects($this->at(2))
+            ->method('log')
+            ->with(
+                $this->equalTo('debug'),
+                $this->equalTo("Emitting aliases via {emitter}"),
+                $this->equalTo(['emitter' => 'Pantheon\Terminus\Helpers\AliasEmitters\AliasesDrushRcEmitter'])
+            );
+        $this->logger->expects($this->at(3))
+            ->method('log')
+            ->with(
+                $this->equalTo('notice'),
+                $this->stringContains(".drush/pantheon.aliases.drushrc.php")
+            );
+        $this->logger->expects($this->at(4))
+            ->method('log')
+            ->with(
+                $this->equalTo('debug'),
+                $this->equalTo("Emitting aliases via {emitter}"),
+                $this->equalTo(['emitter' => 'Pantheon\Terminus\Helpers\AliasEmitters\DrushSitesYmlEmitter'])
+            );
+        $this->logger->expects($this->at(5))
+            ->method('log')
+            ->with(
+                $this->equalTo('notice'),
+                $this->stringContains(".drush/sites/pantheon")
             );
 
         $out = $this->command->aliases();
         $this->assertNull($out);
-    }
 
-    /**
-     * Tests the aliases command when writing to a named file
-     */
-    public function testAliasesWithLocation()
-    {
-        $location = '~/.terminus/behatcache/aliases.php';
+        if (substr(PHP_OS, 0, 3) == 'WIN') {
+            $this->MarkTestSkipped("Temp file handling on Windows is not working correctly in this test.");
+        }
 
-        $this->container->expects($this->once())
-            ->method('get')
-            ->with($this->equalTo(LocalMachineHelper::class))
-            ->willReturn($this->local_machine_helper);
-        $this->local_machine_helper->expects($this->once())
-            ->method('writeFile')
-            ->with(
-                $this->equalTo($location),
-                $this->equalTo($this->aliases)
-            );
-        $this->logger->expects($this->once())
-            ->method('log')
-            ->with(
-                $this->equalTo('notice'),
-                $this->equalTo('Aliases file written to {location}.'),
-                $this->equalTo(compact('location'))
-            );
+        $expected_drush_8_alias_path = $this->home_dir . '/.drush/pantheon.aliases.drushrc.php';
+        $this->assertFileExists($expected_drush_8_alias_path);
+        $drush_8_aliases = file_get_contents($expected_drush_8_alias_path);
 
-        $out = $this->command->aliases(compact('location'));
-        $this->assertNull($out);
+        $this->assertEquals($this->expectedDrush8AliasOutput(), trim($drush_8_aliases));
+
+        $expected_drush_9_alias_path = $this->home_dir . '/.drush/sites/pantheon/site1.site.yml';
+        $this->assertFileExists($expected_drush_9_alias_path);
+        $drush_9_aliases = file_get_contents($expected_drush_9_alias_path);
+        $expected = <<<__EOT__
+'*':
+  host: appserver.\${env-name}.abc.drush.in
+  paths:
+    files: files
+    drush-script: drush9
+  uri: \${env-name}-site1.pantheonsite.io
+  user: \${env-name}.abc
+  ssh:
+    options: '-p 2222 -o "AddressFamily inet"'
+    tty: false
+__EOT__;
+        $this->assertEquals($expected, trim($drush_9_aliases));
     }
 
     /**
@@ -140,14 +168,81 @@ class AliasesCommandTest extends CommandTestCase
      */
     public function testAliasesPrint()
     {
+        $this->logger->expects($this->at(0))
+            ->method('log')
+            ->with(
+                $this->equalTo('notice'),
+                $this->equalTo('Fetching site information to build Drush aliases...')
+            );
+        $this->logger->expects($this->at(1))
+            ->method('log')
+            ->with(
+                $this->equalTo('notice'),
+                $this->equalTo('{count} sites found.'),
+                $this->equalTo(['count' => 1])
+            );
+        $this->logger->expects($this->at(2))
+            ->method('log')
+            ->with(
+                $this->equalTo('debug'),
+                $this->equalTo("Emitting aliases via {emitter}"),
+                $this->equalTo(['emitter' => 'Pantheon\Terminus\Helpers\AliasEmitters\PrintingEmitter'])
+            );
+        $this->logger->expects($this->at(3))
+            ->method('log')
+            ->with(
+                $this->equalTo('notice'),
+                $this->equalTo('Displaying Drush 8 alias file contents.')
+            );
+
         $this->container->expects($this->never())
             ->method('get');
-        $this->local_machine_helper->expects($this->never())
-            ->method('writeFile');
-        $this->logger->expects($this->never())
-            ->method('log');
 
-        $out = $this->command->aliases(['print' => true,]);
-        $this->assertEquals($out, $this->aliases);
+        $out = $this->command->aliases([
+            'print' => true,
+            'location' => null,
+            'all' => false,
+            'only' => '',
+            'type' => 'all',
+            'base' => '~/.drush',
+            'db-url' => true,
+            'target' => 'pantheon',
+        ]);
+        $this->assertNull($out);
+
+        $this->assertEquals($this->expectedDrush8AliasOutput(), trim($this->output->fetch()));
+    }
+
+    /**
+     * Return the expected output for Drush 8 alias files.
+     */
+    protected function expectedDrush8AliasOutput()
+    {
+        $expected = <<<__EOT__
+<?php
+  /**
+   * Pantheon drush alias file, to be placed in your ~/.drush directory or the aliases
+   * directory of your local Drush home. Once it's in place, clear drush cache:
+   *
+   * drush cc drush
+   *
+   * To see all your available aliases:
+   *
+   * drush sa
+   *
+   * See http://helpdesk.getpantheon.com/customer/portal/articles/411388 for details.
+   */
+  \$aliases['site1.*'] = array(
+    'uri' => '\${env-name}-site1.pantheonsite.io',
+    'remote-host' => 'appserver.\${env-name}.abc.drush.in',
+    'remote-user' => '\${env-name}.abc',
+    'ssh-options' => '-p 2222 -o "AddressFamily inet"',
+    'path-aliases' => array(
+      '%files' => 'files',
+      '%drush-script' => 'drush',
+     ),
+  );
+__EOT__;
+        return $expected;
     }
 }